+++ projection/winrt-Windows.ApplicationModel.Store.Preview/py.Windows.ApplicationModel.Store.Preview.cpp
@@ -972,12 +972,18 @@ namespace py::cpp::Windows::ApplicationModel::Store::Preview
         }
     }
 
-    static PyObject* StoreConfiguration_put_PurchasePromptingPolicy(PyObject* /*unused*/, PyObject* arg, void* /*unused*/) noexcept
+    static int StoreConfiguration_put_PurchasePromptingPolicy(PyObject* /*unused*/, PyObject* arg, void* /*unused*/) noexcept
     {
         if (!winrt::Windows::Foundation::Metadata::ApiInformation::IsPropertyPresent(L"Windows.ApplicationModel.Store.Preview.StoreConfiguration", L"PurchasePromptingPolicy"))
         {
             PyErr_SetString(PyExc_AttributeError, "property is not available in this version of Windows");
-            return nullptr;
+            return -1;
+        }
+
+        if (arg == nullptr)
+        {
+            PyErr_SetString(PyExc_AttributeError, "can't delete attribute");
+            return -1;
         }
 
         try
@@ -985,12 +991,12 @@ namespace py::cpp::Windows::ApplicationModel::Store::Preview
             auto param0 = py::convert_to<winrt::Windows::Foundation::IReference<uint32_t>>(arg);
 
             winrt::Windows::ApplicationModel::Store::Preview::StoreConfiguration::PurchasePromptingPolicy(param0);
-            Py_RETURN_NONE;
+            return 0;
         }
         catch (...)
         {
             py::to_PyErr();
-            return nullptr;
+            return -1;
         }
     }