+++ projection/winrt-Windows.Globalization/py.Windows.Globalization.cpp
@@ -65,12 +65,18 @@ namespace py::cpp::Windows::Globalization
         }
     }
 
-    static PyObject* ApplicationLanguages_put_PrimaryLanguageOverride(PyObject* /*unused*/, PyObject* arg, void* /*unused*/) noexcept
+    static int ApplicationLanguages_put_PrimaryLanguageOverride(PyObject* /*unused*/, PyObject* arg, void* /*unused*/) noexcept
     {
         if (!winrt::Windows::Foundation::Metadata::ApiInformation::IsPropertyPresent(L"Windows.Globalization.ApplicationLanguages", L"PrimaryLanguageOverride"))
         {
             PyErr_SetString(PyExc_AttributeError, "property is not available in this version of Windows");
-            return nullptr;
+            return -1;
+        }
+
+        if (arg == nullptr)
+        {
+            PyErr_SetString(PyExc_AttributeError, "can't delete attribute");
+            return -1;
         }
 
         try
@@ -78,12 +84,12 @@ namespace py::cpp::Windows::Globalization
             auto param0 = py::convert_to<winrt::hstring>(arg);
 
             winrt::Windows::Globalization::ApplicationLanguages::PrimaryLanguageOverride(param0);
-            Py_RETURN_NONE;
+            return 0;
         }
         catch (...)
         {
             py::to_PyErr();
-            return nullptr;
+            return -1;
         }
     }