+++ projection/winrt-Windows.Networking.BackgroundTransfer/py.Windows.Networking.BackgroundTransfer.cpp
@@ -2642,12 +2642,18 @@ namespace py::cpp::Windows::Networking::BackgroundTransfer
         }
     }
 
-    static PyObject* ContentPrefetcher_put_IndirectContentUri(PyObject* /*unused*/, PyObject* arg, void* /*unused*/) noexcept
+    static int ContentPrefetcher_put_IndirectContentUri(PyObject* /*unused*/, PyObject* arg, void* /*unused*/) noexcept
     {
         if (!winrt::Windows::Foundation::Metadata::ApiInformation::IsPropertyPresent(L"Windows.Networking.BackgroundTransfer.ContentPrefetcher", L"IndirectContentUri"))
         {
             PyErr_SetString(PyExc_AttributeError, "property is not available in this version of Windows");
-            return nullptr;
+            return -1;
+        }
+
+        if (arg == nullptr)
+        {
+            PyErr_SetString(PyExc_AttributeError, "can't delete attribute");
+            return -1;
         }
 
         try
@@ -2655,12 +2661,12 @@ namespace py::cpp::Windows::Networking::BackgroundTransfer
             auto param0 = py::convert_to<winrt::Windows::Foundation::Uri>(arg);
 
             winrt::Windows::Networking::BackgroundTransfer::ContentPrefetcher::IndirectContentUri(param0);
-            Py_RETURN_NONE;
+            return 0;
         }
         catch (...)
         {
             py::to_PyErr();
-            return nullptr;
+            return -1;
         }
     }