+++ projection/winrt-Windows.UI.Xaml.Media.Animation/py.Windows.UI.Xaml.Media.Animation.cpp
@@ -20150,12 +20150,18 @@ namespace py::cpp::Windows::UI::Xaml::Media::Animation
         }
     }
 
-    static PyObject* Timeline_put_AllowDependentAnimations(PyObject* /*unused*/, PyObject* arg, void* /*unused*/) noexcept
+    static int Timeline_put_AllowDependentAnimations(PyObject* /*unused*/, PyObject* arg, void* /*unused*/) noexcept
     {
         if (!winrt::Windows::Foundation::Metadata::ApiInformation::IsPropertyPresent(L"Windows.UI.Xaml.Media.Animation.Timeline", L"AllowDependentAnimations"))
         {
             PyErr_SetString(PyExc_AttributeError, "property is not available in this version of Windows");
-            return nullptr;
+            return -1;
+        }
+
+        if (arg == nullptr)
+        {
+            PyErr_SetString(PyExc_AttributeError, "can't delete attribute");
+            return -1;
         }
 
         try
@@ -20163,12 +20169,12 @@ namespace py::cpp::Windows::UI::Xaml::Media::Animation
             auto param0 = py::convert_to<bool>(arg);
 
             winrt::Windows::UI::Xaml::Media::Animation::Timeline::AllowDependentAnimations(param0);
-            Py_RETURN_NONE;
+            return 0;
         }
         catch (...)
         {
             py::to_PyErr();
-            return nullptr;
+            return -1;
         }
     }